Changed module name and fixed issues with empty gcode response upon machine startup
